+++ src/backend/avx2/mod.rs
@@ -479,3 +479,5 @@ pub(crate) mod field;
 pub(crate) mod edwards;
 
 pub(crate) mod constants;
+
+pub(crate) mod scalar_mul;
